requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
streamlit>=1.30.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0 